face_blur_api/requirements.txt

Django>=4.2
djangorestframework
opencv-python-headless
numpy
pillow
